+++ packages/v2-sdk/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sushiswap/v2-sdk
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [d6b7f8443]
+  - @sushiswap/chain@0.1.1
+  - @sushiswap/currency@0.0.5
+
 ## 0.0.3
 
 ### Patch Changes